refactor: extract activation trace package
Move Codex-managed runtime trace conversion, redaction, and bounding into a dedicated activationtrace package while keeping the existing driver API as a compatibility facade for current callers.

Validation: go test ./harness/internal/activationtrace ./harness/internal/driver ./harness/cmd/mnemon-multica-runtime
